--- backend/smody/src/main/java/com/woowacourse/smody/push/strategy/PushStrategy.java
@@ -1,13 +1,0 @@
-package com.woowacourse.smody.push.strategy;
-
-import com.woowacourse.smody.push.domain.PushCase;
-import com.woowacourse.smody.push.domain.PushNotification;
-
-public interface PushStrategy {
-
-    void push(Object entity);
-
-    PushCase getPushCase();
-
-    PushNotification buildNotification(Object entity);
-}
