+++ backend/smody/src/test/java/com/woowacourse/smody/push/event/EventExceptionTest.java
@@ -1,13 +1,21 @@
 package com.woowacourse.smody.push.event;
 
-import static com.woowacourse.smody.support.ResourceFixture.더즈_ID;
-import static com.woowacourse.smody.support.ResourceFixture.미라클_모닝_ID;
-import static com.woowacourse.smody.support.ResourceFixture.스모디_방문하기_ID;
-import static com.woowacourse.smody.support.ResourceFixture.조조그린_ID;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.willThrow;
+import static com.woowacourse.smody.support.ResourceFixture.*;
+import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.BDDMockito.*;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 import com.woowacourse.smody.auth.dto.TokenPayload;
 import com.woowacourse.smody.comment.domain.Comment;
@@ -26,16 +34,6 @@ import com.woowacourse.smody.push.dto.SubscriptionRequest;
 import com.woowacourse.smody.push.repository.PushNotificationRepository;
 import com.woowacourse.smody.push.service.PushSubscriptionService;
 import com.woowacourse.smody.support.IntegrationTest;
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
-import java.util.concurrent.TimeUnit;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 @DisplayName("알림 이벤트에 예외가 발생해도 ")
 class EventExceptionTest extends IntegrationTest {
@@ -56,7 +54,13 @@ class EventExceptionTest extends IntegrationTest {
 	private ThreadPoolTaskExecutor executor;
 
 	@MockBean
-	private PushEventListener pushEventListener;
+	private ChallengePushEventListener challengePushStrategy;
+
+	@MockBean
+	private SubscriptionPushEventListener subscriptionPushStrategy;
+
+	@MockBean
+	private CommentPushEventListener commentPushStrategy;
 
 	@DisplayName("새로운 사이클이 저장된다.")
 	@Test
@@ -65,7 +69,7 @@ class EventExceptionTest extends IntegrationTest {
 		LocalDateTime now = LocalDateTime.now();
 
 		willThrow(new RuntimeException("알림 로직에 예상치 못한 예외 발생!"))
-			.given(pushEventListener).handleCycleCreate(any(CycleCreateEvent.class));
+			.given(challengePushStrategy).handle(any(CycleCreateEvent.class));
 
 		// when
 		Long cycleId = cycleService.create(
@@ -93,7 +97,7 @@ class EventExceptionTest extends IntegrationTest {
 			"endpoint-link", "p256dh", "auth");
 
 		willThrow(new RuntimeException("알림 로직에 예상치 못한 예외 발생!"))
-			.given(pushEventListener).handlePushSubscribe(any(PushSubscribeEvent.class));
+			.given(subscriptionPushStrategy).handle(any(PushSubscribeEvent.class));
 
 		// when
 		pushSubscriptionService.subscribe(tokenPayload, subscriptionRequest);
@@ -123,7 +127,7 @@ class EventExceptionTest extends IntegrationTest {
 		CommentRequest commentRequest = new CommentRequest("댓글입니다");
 
 		willThrow(new RuntimeException("알림 로직에 예상치 못한 예외 발생!"))
-			.given(pushEventListener).handleCommentCreate(any(CommentCreateEvent.class));
+			.given(commentPushStrategy).handle(any(CommentCreateEvent.class));
 
 		// when
 		Long commentId = commentService.create(new TokenPayload(더즈_ID), cycleDetail.getId(), commentRequest);
