+++ backend/smody/src/test/java/com/woowacourse/smody/push/event/SubscriptionPushEventListenerTest.java
@@ -4,50 +4,44 @@ import static com.woowacourse.smody.support.ResourceFixture.조조그린_ID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import com.woowacourse.smody.auth.dto.TokenPayload;
+import java.util.concurrent.TimeUnit;
+
 import com.woowacourse.smody.push.domain.PushCase;
 import com.woowacourse.smody.push.domain.PushNotification;
 import com.woowacourse.smody.push.domain.PushStatus;
-import com.woowacourse.smody.push.dto.SubscriptionRequest;
+import com.woowacourse.smody.push.domain.PushSubscribeEvent;
+import com.woowacourse.smody.push.domain.PushSubscription;
 import com.woowacourse.smody.push.repository.PushNotificationRepository;
-import com.woowacourse.smody.push.service.PushSubscriptionService;
 import com.woowacourse.smody.support.IntegrationTest;
-import java.util.concurrent.TimeUnit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
-class SubscriptionPushEventListenerTest extends IntegrationTest {
+public class SubscriptionPushEventListenerTest extends IntegrationTest {
+
+	@Autowired
+	private SubscriptionPushEventListener pushStrategy;
 
 	@Autowired
 	private PushNotificationRepository pushNotificationRepository;
 
-	@Autowired
-	private PushSubscriptionService pushSubscriptionService;
-
-	@Autowired
-	private ThreadPoolTaskExecutor executor;
-
-	@DisplayName("알림을 구독하면 푸시 알람 내역이 발송된 상태로 저장된다.")
+	@DisplayName("알림 구독에 대한 알림을 전송한다.")
 	@Test
-	void subscribe_pushNotification() throws InterruptedException {
+	void push() throws InterruptedException {
 		// given
-		TokenPayload tokenPayload = new TokenPayload(조조그린_ID);
-		SubscriptionRequest subscriptionRequest = new SubscriptionRequest(
-			"endpoint-link", "p256dh", "auth");
+		PushSubscription pushSubscription = fixture.알림_구독(조조그린_ID, "endpoint");
 
 		// when
-		pushSubscriptionService.subscribe(tokenPayload, subscriptionRequest);
+		pushStrategy.handle(new PushSubscribeEvent(pushSubscription));
 
-		executor.getThreadPoolExecutor().awaitTermination(1, TimeUnit.SECONDS);
+		taskExecutor.getThreadPoolExecutor().awaitTermination(1, TimeUnit.SECONDS);
 
 		// then
-		PushNotification pushNotification = pushNotificationRepository.findAll().get(0);
+		PushNotification pushNotification = pushNotificationRepository.findByPushStatus(PushStatus.COMPLETE).get(0);
 		assertAll(
 			() -> assertThat(pushNotification.getMember().getId()).isEqualTo(조조그린_ID),
 			() -> assertThat(pushNotification.getPushStatus()).isEqualTo(PushStatus.COMPLETE),
-			() -> assertThat(pushNotification.getMessage()).isEqualTo("조조그린님 스모디 알림이 구독되었습니다."),
+			() -> assertThat(pushNotification.getMessage()).contains("조조그린님 스모디 알림이 구독되었습니다."),
 			() -> assertThat(pushNotification.getPushCase()).isEqualTo(PushCase.SUBSCRIPTION),
 			() -> assertThat(pushNotification.getPathId()).isNull()
 		);
