+++ backend/smody/src/test/java/com/woowacourse/smody/support/IntegrationTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 @SpringBootTest
 @ExtendWith(SmodyTestEnvironmentExtension.class)
@@ -20,4 +21,7 @@ public class IntegrationTest {
 
     @Autowired
     protected ResourceFixture fixture;
+
+    @Autowired
+    protected ThreadPoolTaskExecutor taskExecutor;
 }
